+++ src/canisters/frontend/ic-certified-assets/src/tests.rs
@@ -2,6 +2,7 @@ use crate::CreateChunksArg;
 use crate::asset_certification::types::http::{
     CallbackFunc, HttpRequest, HttpResponse, StreamingCallbackToken, StreamingStrategy,
 };
+use crate::canister_env::CanisterEnv;
 use crate::state_machine::{BATCH_EXPIRY_NANOS, StableStateV2, State};
 use crate::types::{
     AssetProperties, BatchId, BatchOperation, CommitBatchArguments, CommitProposedBatchArguments,
@@ -9,7 +10,7 @@ use crate::types::{
     DeleteBatchArguments, GetArg, GetChunkArg, SetAssetContentArguments,
     SetAssetPropertiesArguments,
 };
-use crate::url_decode::{UrlDecodeError, url_decode};
+use crate::url::{UrlDecodeError, url_decode};
 use candid::{Nat, Principal};
 use ic_certification_testing::CertificateBuilder;
 use ic_crypto_tree_hash::Digest;
@@ -18,11 +19,14 @@ use ic_response_verification_test_utils::{
     base64_encode, create_canister_id, get_current_timestamp,
 };
 use serde_bytes::ByteBuf;
-use std::collections::HashMap;
+use std::collections::{BTreeMap, HashMap};
 use std::str::FromStr;
 
 // from ic-response-verification tests
 const MAX_CERT_TIME_OFFSET_NS: u128 = 300_000_000_000;
+
+/// The empty canister env value serialized as a cookie value
+const DEFAULT_IC_ENV_COOKIE_VALUE: &str = "ic_env=ic%5Froot%5Fkey%3D; SameSite=Lax";
 
 fn some_principal() -> Principal {
     Principal::from_text("ryjl3-tyaaa-aaaaa-aaaba-cai").unwrap()
@@ -30,6 +34,13 @@ fn some_principal() -> Principal {
 
 fn unused_callback() -> CallbackFunc {
     CallbackFunc::new(some_principal(), "unused".to_string())
+}
+
+fn empty_canister_env() -> CanisterEnv {
+    CanisterEnv {
+        ic_root_key: vec![],
+        icp_public_env_vars: HashMap::new(),
+    }
 }
 
 pub fn verify_response(
@@ -106,7 +117,7 @@ struct AssetBuilder {
     content_type: String,
     encodings: Vec<(String, Vec<ByteBuf>)>,
     max_age: Option<u64>,
-    headers: Option<HashMap<String, String>>,
+    headers: Option<BTreeMap<String, String>>,
     aliasing: Option<bool>,
     allow_raw_access: Option<bool>,
 }
@@ -141,7 +152,7 @@ impl AssetBuilder {
     }
 
     fn with_header(mut self, header_key: &str, header_value: &str) -> Self {
-        let hm = self.headers.get_or_insert(HashMap::new());
+        let hm = self.headers.get_or_insert(BTreeMap::new());
         hm.insert(header_key.to_string(), header_value.to_string());
         self
     }
@@ -198,7 +209,12 @@ impl RequestBuilder {
     }
 }
 
-fn create_assets(state: &mut State, time_now: u64, assets: Vec<AssetBuilder>) -> BatchId {
+fn create_assets(
+    state: &mut State,
+    time_now: u64,
+    canister_env: &CanisterEnv,
+    assets: Vec<AssetBuilder>,
+) -> BatchId {
     let batch_id = state.create_batch(time_now).unwrap();
 
     let operations =
@@ -211,6 +227,7 @@ fn create_assets(state: &mut State, time_now: u64, assets: Vec<AssetBuilder>) ->
                 operations,
             },
             time_now,
+            canister_env,
         )
         .unwrap();
 
@@ -220,6 +237,7 @@ fn create_assets(state: &mut State, time_now: u64, assets: Vec<AssetBuilder>) ->
 fn create_assets_by_proposal(
     state: &mut State,
     time_now: u64,
+    canister_env: &CanisterEnv,
     assets: Vec<AssetBuilder>,
 ) -> BatchId {
     let batch_id = state.create_batch(time_now).unwrap();
@@ -249,6 +267,7 @@ fn create_assets_by_proposal(
                 evidence,
             },
             time_now,
+            canister_env,
         )
         .unwrap();
 
@@ -332,7 +351,7 @@ impl State {
     }
 
     fn create_test_asset(&mut self, asset: AssetBuilder) {
-        create_assets(self, 100_000_000_000, vec![asset]);
+        create_assets(self, 100_000_000_000, &empty_canister_env(), vec![asset]);
     }
 }
 
@@ -340,12 +359,14 @@ impl State {
 fn can_create_assets_using_batch_api() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
     let batch_id = create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html").with_encoding("identity", vec![BODY]),
         ],
@@ -385,6 +406,7 @@ fn can_create_assets_using_batch_api() {
 fn serve_correct_encoding_v1() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const IDENTITY_BODY: &[u8] = b"<!DOCTYPE html><html></html>";
     const GZIP_BODY: &[u8] = b"this is 'gzipped' content";
@@ -392,6 +414,7 @@ fn serve_correct_encoding_v1() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![IDENTITY_BODY])
@@ -463,6 +486,7 @@ fn serve_correct_encoding_v1() {
 fn serve_correct_encoding_v2() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const IDENTITY_BODY: &[u8] = b"<!DOCTYPE html><html></html>";
     const GZIP_BODY: &[u8] = b"this is 'gzipped' content";
@@ -470,6 +494,7 @@ fn serve_correct_encoding_v2() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![IDENTITY_BODY])
@@ -516,6 +541,7 @@ fn serve_correct_encoding_v2() {
 fn serve_fallback_v2() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const INDEX_BODY: &[u8] = b"<!DOCTYPE html><html></html>";
     const OTHER_BODY: &[u8] = b"<!DOCTYPE html><html>other content</html>";
@@ -523,6 +549,7 @@ fn serve_fallback_v2() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/index.html", "text/html")
                 .with_encoding("identity", vec![INDEX_BODY]),
@@ -588,12 +615,14 @@ fn serve_fallback_v2() {
 fn serve_fallback_v1() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const INDEX_BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/index.html", "text/html")
                 .with_encoding("identity", vec![INDEX_BODY]),
@@ -625,12 +654,14 @@ fn serve_fallback_v1() {
 fn can_create_assets_using_batch_proposal_api() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
     let batch_id = create_assets_by_proposal(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html").with_encoding("identity", vec![BODY]),
         ],
@@ -896,6 +927,7 @@ fn can_delete_batch_with_chunks() {
 fn returns_index_file_for_missing_assets() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const INDEX_BODY: &[u8] = b"<!DOCTYPE html><html>Index</html>";
     const OTHER_BODY: &[u8] = b"<!DOCTYPE html><html>Other</html>";
@@ -903,6 +935,7 @@ fn returns_index_file_for_missing_assets() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/index.html", "text/html")
                 .with_encoding("identity", vec![INDEX_BODY]),
@@ -926,12 +959,14 @@ fn returns_index_file_for_missing_assets() {
 fn preserves_state_on_stable_roundtrip() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const INDEX_BODY: &[u8] = b"<!DOCTYPE html><html>Index</html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/index.html", "text/html")
                 .with_encoding("identity", vec![INDEX_BODY]),
@@ -955,6 +990,7 @@ fn preserves_state_on_stable_roundtrip() {
 fn uses_streaming_for_multichunk_assets() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const INDEX_BODY_CHUNK_1: &[u8] = b"<!DOCTYPE html>";
     const INDEX_BODY_CHUNK_2: &[u8] = b"<html>Index</html>";
@@ -962,6 +998,7 @@ fn uses_streaming_for_multichunk_assets() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/index.html", "text/html")
                 .with_encoding("identity", vec![INDEX_BODY_CHUNK_1, INDEX_BODY_CHUNK_2]),
@@ -1011,6 +1048,7 @@ fn uses_streaming_for_multichunk_assets() {
 fn get_and_get_chunk_for_multichunk_assets() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const INDEX_BODY_CHUNK_0: &[u8] = b"<!DOCTYPE html>";
     const INDEX_BODY_CHUNK_1: &[u8] = b"<html>Index</html>";
@@ -1018,6 +1056,7 @@ fn get_and_get_chunk_for_multichunk_assets() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/index.html", "text/html")
                 .with_encoding("identity", vec![INDEX_BODY_CHUNK_0, INDEX_BODY_CHUNK_1]),
@@ -1060,12 +1099,14 @@ fn get_and_get_chunk_for_multichunk_assets() {
 fn supports_max_age_headers() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html").with_encoding("identity", vec![BODY]),
             AssetBuilder::new("/max-age.html", "text/html")
@@ -1140,12 +1181,14 @@ fn check_url_decode() {
 fn supports_custom_http_headers() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![BODY])
@@ -1208,12 +1251,15 @@ fn supports_custom_http_headers() {
 fn supports_getting_and_setting_asset_properties() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
 
     const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
+    let set_cookie_header = ("Set-Cookie".into(), DEFAULT_IC_ENV_COOKIE_VALUE.into());
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![BODY])
@@ -1229,10 +1275,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/contents.html".into()),
         Ok(AssetProperties {
             max_age: None,
-            headers: Some(HashMap::from([(
-                "Access-Control-Allow-Origin".into(),
-                "*".into()
-            )])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("Access-Control-Allow-Origin".into(), "*".into())
+            ])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1241,10 +1287,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(604800),
-            headers: Some(HashMap::from([(
-                "X-Content-Type-Options".into(),
-                "nosniff".into()
-            )])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("X-Content-Type-Options".into(), "nosniff".into())
+            ])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1255,7 +1301,7 @@ fn supports_getting_and_setting_asset_properties() {
             .set_asset_properties(SetAssetPropertiesArguments {
                 key: "/max-age.html".into(),
                 max_age: Some(Some(1)),
-                headers: Some(Some(HashMap::from([(
+                headers: Some(Some(BTreeMap::from([(
                     "X-Content-Type-Options".into(),
                     "nosniff".into()
                 )]))),
@@ -1268,10 +1314,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(1),
-            headers: Some(HashMap::from([(
-                "X-Content-Type-Options".into(),
-                "nosniff".into()
-            )])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("X-Content-Type-Options".into(), "nosniff".into())
+            ])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1292,7 +1338,7 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: None,
-            headers: None,
+            headers: Some(BTreeMap::from([set_cookie_header.clone()])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1303,7 +1349,7 @@ fn supports_getting_and_setting_asset_properties() {
             .set_asset_properties(SetAssetPropertiesArguments {
                 key: "/max-age.html".into(),
                 max_age: Some(Some(1)),
-                headers: Some(Some(HashMap::from([(
+                headers: Some(Some(BTreeMap::from([(
                     "X-Content-Type-Options".into(),
                     "nosniff".into()
                 )]))),
@@ -1316,10 +1362,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(1),
-            headers: Some(HashMap::from([(
-                "X-Content-Type-Options".into(),
-                "nosniff".into()
-            )])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("X-Content-Type-Options".into(), "nosniff".into())
+            ])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1330,7 +1376,10 @@ fn supports_getting_and_setting_asset_properties() {
             .set_asset_properties(SetAssetPropertiesArguments {
                 key: "/max-age.html".into(),
                 max_age: None,
-                headers: Some(Some(HashMap::from([("new-header".into(), "value".into())]))),
+                headers: Some(Some(BTreeMap::from([(
+                    "new-header".into(),
+                    "value".into()
+                )]))),
                 allow_raw_access: None,
                 is_aliased: None
             })
@@ -1340,7 +1389,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(1),
-            headers: Some(HashMap::from([("new-header".into(), "value".into())])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("new-header".into(), "value".into())
+            ])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1361,7 +1413,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(2),
-            headers: Some(HashMap::from([("new-header".into(), "value".into())])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("new-header".into(), "value".into())
+            ])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1382,7 +1437,10 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(2),
-            headers: Some(HashMap::from([("new-header".into(), "value".into())])),
+            headers: Some(BTreeMap::from([
+                set_cookie_header.clone(),
+                ("new-header".into(), "value".into())
+            ])),
             allow_raw_access: None,
             is_aliased: Some(false)
         })
@@ -1403,7 +1461,7 @@ fn supports_getting_and_setting_asset_properties() {
         state.get_asset_properties("/max-age.html".into()),
         Ok(AssetProperties {
             max_age: Some(2),
-            headers: None,
+            headers: Some(BTreeMap::from([set_cookie_header.clone()])),
             allow_raw_access: None,
             is_aliased: None
         })
@@ -1414,11 +1472,13 @@ fn supports_getting_and_setting_asset_properties() {
 fn create_asset_fails_if_asset_exists() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
     const FILE_BODY: &[u8] = b"<!DOCTYPE html><html>file body</html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![FILE_BODY]),
@@ -1444,12 +1504,15 @@ fn create_asset_fails_if_asset_exists() {
 fn support_aliases() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
     const INDEX_BODY: &[u8] = b"<!DOCTYPE html><html>index</html>";
     const SUBDIR_INDEX_BODY: &[u8] = b"<!DOCTYPE html><html>subdir index</html>";
     const FILE_BODY: &[u8] = b"<!DOCTYPE html><html>file body</html>";
+
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![FILE_BODY]),
@@ -1492,12 +1555,14 @@ fn support_aliases() {
 fn alias_enable_and_disable() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
     const SUBDIR_INDEX_BODY: &[u8] = b"<!DOCTYPE html><html>subdir index</html>";
     const FILE_BODY: &[u8] = b"<!DOCTYPE html><html>file body</html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![FILE_BODY]),
@@ -1535,6 +1600,7 @@ fn alias_enable_and_disable() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![FILE_BODY])
@@ -1562,12 +1628,14 @@ fn alias_enable_and_disable() {
 fn alias_behavior_persists_through_upgrade() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
     const SUBDIR_INDEX_BODY: &[u8] = b"<!DOCTYPE html><html>subdir index</html>";
     const FILE_BODY: &[u8] = b"<!DOCTYPE html><html>file body</html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![FILE_BODY])
@@ -1613,12 +1681,14 @@ fn alias_behavior_persists_through_upgrade() {
 fn aliasing_name_clash() {
     let mut state = State::default();
     let time_now = 100_000_000_000;
+    let canister_env = empty_canister_env();
     const FILE_BODY: &[u8] = b"<!DOCTYPE html><html>file body</html>";
     const FILE_BODY_2: &[u8] = b"<!DOCTYPE html><html>second body</html>";
 
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents.html", "text/html")
                 .with_encoding("identity", vec![FILE_BODY]),
@@ -1631,6 +1701,7 @@ fn aliasing_name_clash() {
     create_assets(
         &mut state,
         time_now,
+        &canister_env,
         vec![
             AssetBuilder::new("/contents", "text/html")
                 .with_encoding("identity", vec![FILE_BODY_2]),
@@ -1651,6 +1722,75 @@ fn aliasing_name_clash() {
     let alias_accessible_again =
         certified_http_request(&state, RequestBuilder::get("/contents").build());
     assert_eq!(alias_accessible_again.body.as_ref(), FILE_BODY);
+}
+
+#[test]
+fn headers_cbor_deserialize_from_hashmap_to_btreemap() {
+    // We want to make sure that deserializing from a HashMap to a BTreeMap works
+    // so that frontend canister upgrades don't break
+    for i in 0..100 {
+        let old_headers: HashMap<String, String> = HashMap::from([
+            // Order is not alphabetical on purpose here
+            // to check that the BTreeMap orders them correctly
+            ("c-name".into(), "c-value".into()),
+            ("index".into(), i.to_string()),
+            ("d-name".into(), "d-value".into()),
+            ("b-name".into(), "b-value".into()),
+            ("a-name".into(), "a-value".into()),
+        ]);
+        let serialized = serde_cbor::to_vec(&old_headers).unwrap();
+        let new_headers: BTreeMap<String, String> = serde_cbor::from_slice(&serialized).unwrap();
+        // Compare the order to check that the BTreeMap is deterministic
+        assert_eq!(
+            new_headers.into_iter().collect::<Vec<(String, String)>>(),
+            vec![
+                ("a-name".into(), "a-value".into()),
+                ("b-name".into(), "b-value".into()),
+                ("c-name".into(), "c-value".into()),
+                ("d-name".into(), "d-value".into()),
+                ("index".into(), i.to_string()),
+            ]
+        );
+    }
+}
+
+#[test]
+fn headers_candid_hashmap_btreemap_roundtrip() {
+    for i in 0..100 {
+        let old_headers: HashMap<String, String> = HashMap::from([
+            ("a-name".into(), "a-value".into()),
+            ("b-name".into(), "b-value".into()),
+            ("c-name".into(), "c-value".into()),
+            ("d-name".into(), "d-value".into()),
+            ("index".into(), i.to_string()),
+        ]);
+
+        // Deserialize to BTreeMap
+        let old_serialized = candid::encode_one(&old_headers).unwrap();
+        let new_headers: BTreeMap<String, String> = candid::decode_one(&old_serialized).unwrap();
+        assert_eq!(
+            new_headers
+                .clone()
+                .into_iter()
+                .collect::<Vec<(String, String)>>(),
+            vec![
+                ("a-name".into(), "a-value".into()),
+                ("b-name".into(), "b-value".into()),
+                ("c-name".into(), "c-value".into()),
+                ("d-name".into(), "d-value".into()),
+                ("index".into(), i.to_string()),
+            ]
+        );
+
+        // Go back to HashMap
+        let new_serialized = candid::encode_one(new_headers).unwrap();
+        let old_deserialized: HashMap<String, String> =
+            candid::decode_one(&new_serialized).unwrap();
+        assert_eq!(
+            old_deserialized, old_headers,
+            "Old headers don't match, iteration: {i}",
+        );
+    }
 }
 
 #[cfg(test)]
@@ -1780,12 +1920,14 @@ mod certificate_expression {
     fn ic_certificate_expression_present_for_new_assets() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
 
         const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
         create_assets(
             &mut state,
             time_now,
+            &canister_env,
             vec![
                 AssetBuilder::new("/contents.html", "text/html")
                     .with_encoding("identity", vec![BODY])
@@ -1821,7 +1963,7 @@ mod certificate_expression {
         );
         assert_eq!(
             lookup_header(&response, "ic-certificateexpression").unwrap(),
-            r#"default_certification(ValidationArgs{certification: Certification{no_request_certification: Empty{}, response_certification: ResponseCertification{certified_response_headers: ResponseHeaderList{headers: ["content-type", "cache-control", "Access-Control-Allow-Origin"]}}}})"#,
+            r#"default_certification(ValidationArgs{certification: Certification{no_request_certification: Empty{}, response_certification: ResponseCertification{certified_response_headers: ResponseHeaderList{headers: ["content-type", "cache-control", "Access-Control-Allow-Origin", "Set-Cookie"]}}}})"#,
             "Missing ic-certifiedexpression header in response: {:#?}",
             response,
         );
@@ -1831,12 +1973,14 @@ mod certificate_expression {
     fn ic_certificate_expression_gets_updated_on_asset_properties_update() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
 
         const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
         create_assets(
             &mut state,
             time_now,
+            &canister_env,
             vec![
                 AssetBuilder::new("/contents.html", "text/html")
                     .with_encoding("gzip", vec![BODY])
@@ -1860,7 +2004,7 @@ mod certificate_expression {
         );
         assert_eq!(
             lookup_header(&response, "ic-certificateexpression").unwrap(),
-            r#"default_certification(ValidationArgs{certification: Certification{no_request_certification: Empty{}, response_certification: ResponseCertification{certified_response_headers: ResponseHeaderList{headers: ["content-type", "content-encoding", "cache-control", "Access-Control-Allow-Origin"]}}}})"#,
+            r#"default_certification(ValidationArgs{certification: Certification{no_request_certification: Empty{}, response_certification: ResponseCertification{certified_response_headers: ResponseHeaderList{headers: ["content-type", "content-encoding", "cache-control", "Access-Control-Allow-Origin", "Set-Cookie"]}}}})"#,
             "Missing ic-certificateexpression header in response: {:#?}",
             response,
         );
@@ -1869,7 +2013,7 @@ mod certificate_expression {
             .set_asset_properties(SetAssetPropertiesArguments {
                 key: "/contents.html".into(),
                 max_age: Some(None),
-                headers: Some(Some(HashMap::from([(
+                headers: Some(Some(BTreeMap::from([(
                     "custom-header".into(),
                     "value".into(),
                 )]))),
@@ -1891,7 +2035,7 @@ mod certificate_expression {
         );
         assert_eq!(
             lookup_header(&response, "ic-certificateexpression").unwrap(),
-            r#"default_certification(ValidationArgs{certification: Certification{no_request_certification: Empty{}, response_certification: ResponseCertification{certified_response_headers: ResponseHeaderList{headers: ["content-type", "content-encoding", "custom-header"]}}}})"#,
+            r#"default_certification(ValidationArgs{certification: Certification{no_request_certification: Empty{}, response_certification: ResponseCertification{certified_response_headers: ResponseHeaderList{headers: ["content-type", "content-encoding", "Set-Cookie", "custom-header"]}}}})"#,
             "Missing ic-certifiedexpression header in response: {:#?}",
             response,
         );
@@ -1906,6 +2050,7 @@ mod certification_v2 {
     fn proper_header_structure() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
 
         const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
         const UPDATED_BODY: &[u8] = b"<!DOCTYPE html><html>lots of content!</html>";
@@ -1913,6 +2058,7 @@ mod certification_v2 {
         create_assets(
             &mut state,
             time_now,
+            &canister_env,
             vec![
                 AssetBuilder::new("/contents.html", "text/html")
                     .with_encoding("identity", vec![BODY])
@@ -1945,6 +2091,7 @@ mod certification_v2 {
         create_assets(
             &mut state,
             time_now,
+            &canister_env,
             vec![
                 AssetBuilder::new("/contents.html", "text/html")
                     .with_encoding("identity", vec![UPDATED_BODY])
@@ -1971,12 +2118,14 @@ mod certification_v2 {
 
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
 
         const BODY: &[u8] = b"<!DOCTYPE html><html></html>";
 
         create_assets(
             &mut state,
             time_now,
+            &canister_env,
             vec![
                 AssetBuilder::new("/contents.html", "text/html")
                     .with_encoding("identity", vec![BODY])
@@ -2012,6 +2161,8 @@ mod certification_v2 {
 
 #[cfg(test)]
 mod evidence_computation {
+    use std::collections::BTreeMap;
+
     use super::*;
     use crate::types::BatchOperation::SetAssetContent;
     use crate::types::{ClearArguments, ComputeEvidenceArguments, UnsetAssetContentArguments};
@@ -2314,7 +2465,7 @@ mod evidence_computation {
                             key: "/d".to_string(),
                             content_type: "text/plain".to_string(),
                             max_age: Some(98),
-                            headers: Some(HashMap::from([
+                            headers: Some(BTreeMap::from([
                                 ("H1".to_string(), "V1".to_string()),
                                 ("H2".to_string(), "V2".to_string())
                             ])),
@@ -2342,7 +2493,7 @@ mod evidence_computation {
                             key: "/d".to_string(),
                             content_type: "text/plain".to_string(),
                             max_age: Some(98),
-                            headers: Some(HashMap::from([
+                            headers: Some(BTreeMap::from([
                                 ("H1".to_string(), "V1".to_string()),
                                 ("H2".to_string(), "V2".to_string())
                             ])),
@@ -2576,7 +2727,7 @@ mod evidence_computation {
                         key: "/".to_string(),
                         content_type: "".to_string(),
                         max_age: None,
-                        headers: Some(HashMap::from([("H1".to_string(), "V1".to_string()),])),
+                        headers: Some(BTreeMap::from([("H1".to_string(), "V1".to_string()),])),
                         enable_aliasing: None,
                         allow_raw_access: None,
                     }),],
@@ -2601,7 +2752,7 @@ mod evidence_computation {
                         key: "/".to_string(),
                         content_type: "".to_string(),
                         max_age: None,
-                        headers: Some(HashMap::from([("H1".to_string(), "V2".to_string()),])),
+                        headers: Some(BTreeMap::from([("H1".to_string(), "V2".to_string()),])),
                         enable_aliasing: None,
                         allow_raw_access: None,
                     }),],
@@ -2626,7 +2777,7 @@ mod evidence_computation {
                         key: "/".to_string(),
                         content_type: "".to_string(),
                         max_age: None,
-                        headers: Some(HashMap::from([("H2".to_string(), "V1".to_string()),])),
+                        headers: Some(BTreeMap::from([("H2".to_string(), "V1".to_string()),])),
                         enable_aliasing: None,
                         allow_raw_access: None,
                     }),],
@@ -2652,7 +2803,7 @@ mod evidence_computation {
                         key: "/".to_string(),
                         content_type: "".to_string(),
                         max_age: None,
-                        headers: Some(HashMap::from([
+                        headers: Some(BTreeMap::from([
                             ("H1".to_string(), "V1".to_string()),
                             ("H2".to_string(), "V2".to_string()),
                         ])),
@@ -3458,7 +3609,7 @@ mod evidence_computation {
                 for headers in &[
                     None,
                     Some(None),
-                    Some(Some(HashMap::from([(
+                    Some(Some(BTreeMap::from([(
                         String::from("a"),
                         String::from("b"),
                     )]))),
@@ -3528,6 +3679,7 @@ mod validate_commit_proposed_batch {
     fn batch_not_found() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
 
         match state.validate_commit_proposed_batch(CommitProposedBatchArguments {
             batch_id: 1_u8.into(),
@@ -3543,6 +3695,7 @@ mod validate_commit_proposed_batch {
                 evidence: Default::default(),
             },
             time_now,
+            &canister_env,
         ) {
             Err(err) if err.contains("batch not found") => (),
             other => panic!("expected 'batch not found' error, got: {:?}", other),
@@ -3553,6 +3706,7 @@ mod validate_commit_proposed_batch {
     fn no_commit_batch_arguments() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
         let batch_id = state.create_batch(time_now).unwrap();
 
         match state.validate_commit_proposed_batch(CommitProposedBatchArguments {
@@ -3569,6 +3723,7 @@ mod validate_commit_proposed_batch {
                 evidence: Default::default(),
             },
             time_now,
+            &canister_env,
         ) {
             Err(err) if err.contains("batch does not have CommitBatchArguments") => (),
             other => panic!("expected 'batch not found' error, got: {:?}", other),
@@ -3579,6 +3734,7 @@ mod validate_commit_proposed_batch {
     fn evidence_not_computed() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
         let batch_id = state.create_batch(time_now).unwrap();
 
         assert!(
@@ -3603,6 +3759,7 @@ mod validate_commit_proposed_batch {
                 evidence: Default::default(),
             },
             time_now,
+            &canister_env,
         ) {
             Err(err) if err.contains("batch does not have computed evidence") => (),
             other => panic!("expected 'batch not found' error, got: {:?}", other),
@@ -3613,6 +3770,7 @@ mod validate_commit_proposed_batch {
     fn evidence_does_not_match() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
         let batch_id = state.create_batch(time_now).unwrap();
 
         assert!(
@@ -3646,6 +3804,7 @@ mod validate_commit_proposed_batch {
                 evidence: Default::default(),
             },
             time_now,
+            &canister_env,
         ) {
             Err(err) if err.contains("does not match presented evidence") => (),
             other => panic!("expected 'batch not found' error, got: {:?}", other),
@@ -3656,6 +3815,7 @@ mod validate_commit_proposed_batch {
     fn all_good() {
         let mut state = State::default();
         let time_now = 100_000_000_000;
+        let canister_env = empty_canister_env();
         let batch_id = state.create_batch(time_now).unwrap();
 
         assert!(
@@ -3693,6 +3853,7 @@ mod validate_commit_proposed_batch {
             .commit_proposed_batch(
                 CommitProposedBatchArguments { batch_id, evidence },
                 time_now,
+                &canister_env,
             )
             .unwrap();
     }