+++ CHANGELOG.md
@@ -16,6 +16,17 @@ dfx ledger --identity anonymous transfer --memo 1 --icp 1000000 "$YOUR_ACCOUNT_I
 dfx cycles convert --amount 100
 ```
 
+### Frontend canister
+
+Use `BTreeMap` instead of `HashMap` for headers to guarantee deterministic ordering.
+
+Sets the `ic_env` cookie for html files, which contains the root key and the canister environment variables that are prefixed with `PUBLIC_`.
+Please note that this version of the frontend canister is only compatible with PocketIC **v10** and above.
+
+- Module hash: 2c5ab2cdebe93356e319b36c33abcb49c2162b97353476bb5894540e6e616f13
+- https://github.com/dfinity/sdk/pull/4392
+- https://github.com/dfinity/sdk/pull/4387
+
 # 0.29.2
 
 ### Frontend canister